requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cssselect==1.2.0
cssutils==2.11.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.2.0
Django==4.2.13
django-bootstrap-v5==1.0.11
django-bootstrap5==24.2
django-crispy-forms==2.1
django-import-export==4.1.1
django-multiselectfield==0.1.13
django-phonenumber-field==7.3.0
django-tenants==3.7.0
et-xmlfile==1.1.0
gunicorn==22.0.0
idna==3.8
lxml==5.3.0
MarkupPy==1.14
more-itertools==10.4.0
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
phonenumbers==8.13.36
pillow==10.3.0
premailer==3.10.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tablib==3.5.0
tabulate==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
